Move CameraPreview to the newly created Camera package
This is the first step in a larger effort to move all camera related
classes to their own package. The reason I want this is for better
code organization and better encapsulation.
